🎨 organizando código
