+++ src/routes/appDrawerScreen.routes.js
@@ -14,7 +14,7 @@ import {
 } from '~/pages/FaleConoscoScreen/tiposDeOcorrencia';
 import PerfilScreen from '~/pages/Perfil';
 import ContaExcluida from '~/pages/Perfil/contaExcluida';
-import ExcluirPerfil from '~/pages/Perfil/excluirPerfil';
+import ExcluirPerfil from '~/pages/Perfil/ExcluirPerfil';
 import PoliticaDePrivacidadeScreen from '~/pages/Perfil/PoliticaDePrivacidade';
 import TermoDeUsoScreen from '~/pages/Perfil/TermosDeUso';
 import SusNoCearaScreen from '~/pages/SusNoCeara';
